… doc/classes/RenderingDevice.xml
… drivers/d3d12/rendering_device_driver_d3d12.cpp
… drivers/metal/metal_objects.mm
… drivers/metal/rendering_device_driver_metal.mm
… drivers/vulkan/rendering_device_driver_vulkan.cpp
… editor/export/shader_baker_export_plugin.cpp
… servers/rendering/multi_uma_buffer.h
… servers/rendering/renderer_rd/forward_clustered/render_forward_clustered.cpp
… servers/rendering/renderer_rd/forward_clustered/render_forward_clustered.h
… servers/rendering/renderer_rd/forward_mobile/render_forward_mobile.cpp
… servers/rendering/renderer_rd/forward_mobile/render_forward_mobile.h
… servers/rendering/renderer_rd/forward_mobile/scene_shader_forward_mobile.cpp
… servers/rendering/renderer_rd/renderer_canvas_render_rd.cpp
… servers/rendering/renderer_rd/renderer_canvas_render_rd.h
… servers/rendering/renderer_rd/shader_rd.cpp
… servers/rendering/renderer_rd/shader_rd.h
… servers/rendering/renderer_rd/storage_rd/render_scene_data_rd.h
… servers/rendering/rendering_device.cpp
… servers/rendering/rendering_device.h
… servers/rendering/rendering_device_commons.h
… servers/rendering/rendering_device_driver.h
… servers/rendering/rendering_device_graph.cpp
… servers/rendering/rendering_device_graph.h
… servers/rendering/rendering_shader_container.cpp
… servers/rendering/rendering_shader_container.h
… thirdparty/d3d12ma/D3D12MemAlloc.cpp
… thirdparty/d3d12ma/D3D12MemAlloc.h
… thirdparty/d3d12ma/LICENSE.txt
… thirdparty/d3d12ma/README.md
… thirdparty/d3d12ma/patches/0001-mingw-support.patch