This work is a heavily refactored and rewritten from TheForge's initial code.

TheForge's original code had too many race conditions and was fundamentally flawed as it was too easy to incur into those data races by accident.

However they identified the proper places that needed changes, and the idea was sound. I used their work as a blueprint to design this work.

This PR implements:

 - A refactor of some shaders to avoid using multiple sets.
 - A push constant emulator.
- Introduction of UMA buffers used by push constants and a few buffers (most notably the ones filled by _fill_instance_data).

Ironically this change seems to positively affect PC more than it does on Mobile:

|             | Before (ms) | After (ms)  | Improvement |
|-------------|--------|--------|-------------|
| Workstation | 0,622  | 0,606  | 2,60 %      |
| Laptop      | 7,326  | 7,220  | 1,44 %      |
| Adreno 640  | 70,833 | 70,819 | 0,02 %      |
| Mali-G68       | 48,894 | 49,138 | -0,50 %     |

**Notes:**

1. This is the [modified TPS Demo](https://github.com/godotengine/tps-demo)
2. This is the Mobile backend. However improvements apply to Clustered, but it's harder to see them because Clustered is much more GPU bound than Mobile.
3. Some of these numbers are unimpressive because as a result of this PR, https://github.com/godotengine/godot/pull/103531 was submitted. This made "before" faster as the bulk of performance difference is in `_fill_instance_data`. The more objects are on scene, the bigger the impact.
4. For some reason before rebasing Adreno 640 "after" got 67.931ms, making it 4.09% faster. But I can no longer reproduce those results. Maybe it was a fluke?
5. I have no idea why Mali-G68 resulted in a performance degradation. It makes no sense. The best explanation is that Mali GPU does not like the Memory Type/Heap the data is being placed on.

**Specs:**

**Workstation:** Ryzen 5900X 2x16GB AMD Radeon 6800 XT 16GB. Ubuntu 24.04 LTS, RADV Mesa 24.3.4
**Laptop:** Ryzen 5700U 4+8GB AMD Radeon Vega 8. Ubuntu 24.04 LTS, RADV Mesa 24.2.8
**Adreno 640:** POCO F2 Pro, stock Android 13
**Mali-G68:** Samsung A54 5G, stock Android 14