@AThousandShips 2025-10-26
Working on pinning down the specifics but this caused a regression on (at least) Windows, seemingly with combinations of parallax layers and ui, seems to be compiler specific

Will make a report on Monday or when I can pin down the specifics